Implemented filtering the item recyclerview

+++ app/src/main/java/com/example/mi4/ui/items/list/ItemRecyclerAdapter.kt
@@ -1,8 +1,13 @@
 package com.example.mi4.ui.items.list
 
+import android.content.Context
+import android.opengl.Visibility
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Filter
+import android.widget.Filterable
+import androidx.core.view.isVisible
 import androidx.recyclerview.widget.RecyclerView
 import com.example.mi4.R
 import com.example.mi4.data.model.Item
@@ -20,8 +25,39 @@ import kotlinx.coroutines.launch
  * channel: Let's Build that app
  * video title: Kotlin Youtube: Intro to RecyclerView (Ep 1)
  */
-class ItemRecyclerAdapter(val itemlist : MutableList<Item>) : RecyclerView.Adapter<CustomViewHolder>() {
+class ItemRecyclerAdapter(private val context: Context, val itemlist : MutableList<Item>) : RecyclerView.Adapter<CustomViewHolder>(), Filterable {
 
+    var itemSearchList: List<Item>? = null
+
+    override fun getFilter(): Filter {
+        return object : Filter() {
+            override fun performFiltering(charSequence: CharSequence): Filter.FilterResults {
+                val charString = charSequence.toString()
+                if (charString.isEmpty()) {
+                   itemSearchList= itemlist
+                } else {
+                    val filteredList = ArrayList<Item>()
+                    for (row in itemlist) {
+                        if (row.naam.toLowerCase().contains(charString.toLowerCase()) || row.kamer.contains(charSequence) || row.type.contains(charSequence)) {
+                            filteredList.add(row)
+                        }
+                    }
+                   itemSearchList= filteredList
+                }
+                val filterResults = Filter.FilterResults()
+                filterResults.values = itemSearchList
+                return filterResults
+            }
+            override fun publishResults(charSequence: CharSequence, filterResults: Filter.FilterResults) {
+               itemSearchList= filterResults.values as ArrayList<Item>
+                notifyDataSetChanged()
+            }
+        }
+    }
+
+    init {
+        itemSearchList = itemlist
+    }
     override fun getItemCount(): Int {
         return itemlist.size
     }
@@ -33,19 +69,26 @@ class ItemRecyclerAdapter(val itemlist : MutableList<Item>) : RecyclerView.Adapt
     }
 
     override fun onBindViewHolder(holder: CustomViewHolder, position: Int) {
-        val item = itemlist[position]
-        holder.view.itemName.text = item.naam
-        holder.view.itemRoom.text = item.kamer
-        holder.view.itemType.text = item.type
-        holder.view.itemValue.text = item.waarde.toString()
-        holder.view.ib_deleteitem.setOnClickListener {
-            GlobalScope.launch{
-                ItemRepositoryImpl().deleteItem(item)
+        if(!(position == itemSearchList!!.size) && !(position > itemSearchList!!.size)){
+            val item = itemSearchList!![position]
+            holder.view.itemName.text = item.naam
+            holder.view.itemRoom.text = item.kamer
+            holder.view.itemType.text = item.type
+            holder.view.itemValue.text = item.waarde.toString()
+            holder.view.ib_deleteitem.setOnClickListener {
+                GlobalScope.launch{
+                    ItemRepositoryImpl().deleteItem(item)
+                }
+                itemlist.removeAt(position)
+                notifyItemRemoved(position)
+                notifyItemRangeChanged(position, itemlist.size)
             }
-            itemlist.removeAt(position)
-            notifyItemRemoved(position)
-            notifyItemRangeChanged(position, itemlist.size)
+            holder.view.isVisible = true // re-show previously hidden items
         }
+        else{
+            holder.view.isVisible = false // hide items which were filtered-out
+        }
+
 
     }
